Added answer to challenge 2

+++ ProjectEuler/Challenge.cs
@@ -3,7 +3,7 @@
     public class Challenge
     {
         public string Name { get; set; }
-        public int Inputs { get; set; }
+        public long Inputs { get; set; }
         public long ExpectedOutput { get; set; }
     }
 }
